added photo preview and picker to screen
